+++ decidim-conferences/app/helpers/decidim/conferences/conferences_helper.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Conferences
+    # Helpers related to the Conferences layout.
+    module ConferencesHelper
+      include Decidim::ResourceHelper
+    end
+  end
+end