+++ decidim-conferences/lib/decidim/conferences/participatory_space.rb
@@ -133,12 +133,10 @@ Decidim.register_participatory_space(:conferences) do |participatory_space|
 
       5.times do
         Decidim::ConferenceSpeaker.create!(
-          user: conference.organization.users.first,
+          user: conference.organization.users.sample,
           full_name: Faker::Name.name,
           position: Decidim::Faker::Localized.word,
-          affiliation: Decidim::Faker::Localized.wrapped("<p>", "</p>") do
-            Decidim::Faker::Localized.paragraph(3)
-          end,
+          affiliation: Decidim::Faker::Localized.paragraph(3),
           short_bio: Decidim::Faker::Localized.wrapped("<p>", "</p>") do
             Decidim::Faker::Localized.paragraph(3)
           end,