+++ decidim-conferences/spec/commands/update_conference_speaker_spec.rb
@@ -10,6 +10,18 @@ module Decidim::Conferences
     let(:conference_speaker) { create :conference_speaker, :with_user, conference: conference }
     let!(:current_user) { create :user, :confirmed, organization: conference.organization }
     let(:user) { nil }
+    let!(:meeting_component) do
+      create(:component, manifest_name: :meetings, participatory_space: conference)
+    end
+    let!(:meetings) do
+      create_list(
+        :meeting,
+        3,
+        component: meeting_component
+      )
+    end
+    let(:meeting_ids) { meetings.map(&:id) }
+
     let(:form) do
       instance_double(
         Admin::ConferenceSpeakerForm,
@@ -23,7 +35,8 @@ module Decidim::Conferences
           affiliation: { en: "new affiliation" },
           short_bio: Decidim::Faker::Localized.sentence(5),
           twitter_handle: "full_name",
-          personal_url: Faker::Internet.url
+          personal_url: Faker::Internet.url,
+          meeting_ids: meeting_ids
         }
       )
     end
@@ -57,6 +70,18 @@ module Decidim::Conferences
         expect { subject.call }.to change(Decidim::ActionLog, :count)
         action_log = Decidim::ActionLog.last
         expect(action_log.version).to be_present
+      end
+
+      it "links meetings" do
+        subject.call
+
+        conference_meetings = []
+        meetings.each do |meeting|
+          conference_meetings << meeting.becomes(Decidim::ConferenceMeeting)
+        end
+
+        conference_speaker.conference_meetings = conference_meetings
+        expect(conference_speaker.conference_meetings).to match_array(conference_meetings)
       end
 
       context "when is an existing user in the platform" do